REview comments

+++ contracts/mock/MockBaseContract.sol
@@ -45,25 +45,25 @@ contract MockBaseHTS is IBaseHTS {
 
     function mintTokenPublic(
         address tokenAddress,
-        int256 amount
+        uint256 amount
     ) external view override returns (int256, int64) {
         string memory tokenName = ERC20Mock(tokenAddress).name();
         if (isFailureToken(tokenName)) {
-            return (23, int64(amount));
+            return (23, int64(int256(amount)));
         } else {
-            return (22, int64(amount));
+            return (22, int64(int256(amount)));
         }
     }
 
     function burnTokenPublic(
         address tokenAddress,
-        int256 amount
+        uint256 amount
     ) external view override returns (int256, int64) {
         string memory tokenName = ERC20Mock(tokenAddress).name();
         if (isFailureToken(tokenName)) {
-            return (23, int64(amount));
+            return (23, int64(int256(amount)));
         } else {
-            return (22, int64(amount));
+            return (22, int64(int256(amount)));
         }
     }
